requirements.txt
numpy==2.1.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
sounddevice==0.5.0
soundfile==0.12.1
